Updated MoveAction implementations to inherit from TurnAction
