+++ src/main/kotlin/io/hsar/wh40k/combatsimulator/logic/TurnAction.kt
@@ -1,6 +1,8 @@
 package io.hsar.wh40k.combatsimulator.logic
 
 import io.hsar.wh40k.combatsimulator.logic.ActionCost.HALF_ACTION
+import io.hsar.wh40k.combatsimulator.model.MapPosition
+import io.hsar.wh40k.combatsimulator.model.MoveAction
 
 sealed class TurnAction {
     abstract val actionCost: ActionCost
@@ -24,6 +26,26 @@ sealed class TurnAction {
     }
 
     data class RELOAD_WEAPON(override val actionCost: ActionCost) : TurnAction()
+
+    object HalfMoveAction: TurnAction(), MoveAction {
+        override val actionCost: ActionCost = ActionCost.HALF_ACTION
+        override fun getMovementRange(agilityBonus: Short): Short {
+            return agilityBonus
+        }
+        override fun isValidMovementPath(startPoint: MapPosition, endPoint: MapPosition): Boolean {
+            return true  // unlike charge etc, there are no special restrictions on half move pathing
+        }
+    }
+
+    object ChargeAction: TurnAction(), MoveAction {
+        override val actionCost: ActionCost = ActionCost.FULL_ACTION
+        override fun getMovementRange(agilityBonus: Short): Short {
+            return (3 * agilityBonus).toShort()
+        }
+        override fun isValidMovementPath(startPoint: MapPosition, endPoint: MapPosition): Boolean {
+            return startPoint - endPoint >= 4
+        }
+    }
 }
 
 interface DamageCausingAction {
