+++ src/test/kotlin/TestWorld.kt
@@ -1,4 +1,4 @@
-import io.hsar.wh40k.combatsimulator.model.HalfMoveAction
+import io.hsar.wh40k.combatsimulator.logic.TurnAction
 import io.hsar.wh40k.combatsimulator.model.MapPosition
 import org.hamcrest.CoreMatchers.equalTo
 import org.hamcrest.MatcherAssert.assertThat
@@ -10,13 +10,13 @@ class TestWorld {
     @Test
     fun testCanMoveToUnitInRange() {
         val world = TestUtils.getGenericTwoUnitWorld(MapPosition(1,6), MapPosition(5, 4))
-        assertThat(world.canMoveToUnit(world.friendlyForces[0], world.enemyForces[0], HalfMoveAction), equalTo(true))
+        assertThat(world.canMoveToUnit(world.friendlyForces[0], world.enemyForces[0], TurnAction.HalfMoveAction), equalTo(true))
     }
 
     @Test
     fun testCanMoveToUnitOutOfRange() {
         val world = TestUtils.getGenericTwoUnitWorld(MapPosition(1,6), MapPosition(6, 4))
-        assertThat(world.canMoveToUnit(world.friendlyForces[0], world.enemyForces[0], HalfMoveAction), equalTo(false))
+        assertThat(world.canMoveToUnit(world.friendlyForces[0], world.enemyForces[0], TurnAction.HalfMoveAction), equalTo(false))
     }
 
     @Test
